requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Reddit scraping
praw==7.7.1

# OpenAI API
openai==1.10.0

# HTTP client for Gumroad
httpx==0.26.0
requests==2.31.0

# Database
sqlalchemy==2.0.25

# Environment management
python-dotenv==1.0.0

# Scheduling
apscheduler==3.10.4

# Logging
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
